+++ src/vm/property.rs
@@ -2582,6 +2582,9 @@ impl Vm {
         for v in self.realm_function_prototypes.values() {
             Self::push_value_roots(&mut roots, v);
         }
+        for v in self.realm_async_function_prototypes.values() {
+            Self::push_value_roots(&mut roots, v);
+        }
         for v in self.realm_array_buffer_prototypes.values() {
             Self::push_value_roots(&mut roots, v);
         }